Several sleepless nights and restructurings later...
Note that the previously unproven 2-out-of-3 lemma was *wrong*.
This is because the Graduate Lemma fails for strong homotopy equivalences as the triangle law there is an equality, not a homotopy.
As a consequence, the step "A -> A^I trivial cofibration for fibrant A" implies weak factorization had to be completely revamped: we now follow properly the route of Gambino and Garner (2008), but since
we a priori only have path objects in the global context, we must make certain ad hoc modifications (reducing to path objects over path objects).

Someone should check all of this!