Implemented improvements for the section on problems of non-cartesian coherence suggested by Nicola. In detail, the corrections and fixes were as follows.
—
P1, L7: replace with "consider arrows j : A —> B and k : C —> D in I and a commutative square (not
necessarily a pullback)…”. Then write the pullback diagram separately.

-- Done.

P1, L8: add justifications: A —> C’ is mono because j is mono, and C’ —> B is mono since it is pullback of
the mono k.

-- Done.

P1, L9: replace with “may not be invertible"

-- It won't be invertible since we explicitly assumed the original square to not be cartesian.
-- However, being non-cartesian is not a "closed" property, so I rephrased in a better way.

P1, L10: I think this diagram (and possibly many others) need numbering for later reference

-- Done.

P1, L15: I would rephrase “It follows that …” by first constructing the diagram and then stating that d is
a diagonal filler for it.

-- Done. This construction is now much more verbose.

P1, L16: I would add the explanation of why the top triangle commutes (using the pullback square with k and
k’ and the coherence of d and e), and then say that the commutativity of the bottom triangle is immediate.
[I generally find it useful to distinguish with some text the parts of the proof that are trivial from
those that are slightly less trivial]

-- Not sure what you meant here, can you check whether this comment still applies after this elaboration?

P1, L18 and below: using “above” twice is somehow confusing, since I imagine you refer to two different
diagrams. Also, it is not clear to me how to construct the factoring of j —> p
through j —> k, in part because I am not sure about what is being assumed at this stage (maybe you can add
some sentences like “Let us now fix …” here and there).

-- The assumptions should be more clear now. Also, the construction of the factorization is depicted.

P1, L-1: I would prefer to write “restricted to A along j : A —> B” and “restricted to C’ along k’ : C’ —>
B”.

P1, L-1: transitivity of what? (I suspect of “agreeing”, but perhaps an alternative phrasing would be
better).

-- Rephrased.

P2, L1&2: I would rephrase by saying “the image of B —> D to lie in the image of “k : C —> D”.

-- Done.

P2, L3: add explanations of the uniqueness: my understanding is that any two fillers are forced to be equal
to C’ —> X. Here, one may naively object that this situation is very special, and in such cases maybe a
unique filler is unavoidable. Perhaps an example would illustrate the problem effectively.

-- Put more stress on the conclusion no longer talking about the coherent fillers for p from awfs in any way.
-- Added the example (morphisms of horn inclusions) that was the starting point of investigation, showing
how badly things break (fibrant objects become discrete).