*** Expanded on connections
We should use a definition of connections that relates to the literature.
I have put that definition in and renamed our old notion "effective connections".
There is now a new proposition showing that the usual notion of connections imply effective connections.

The new notion of connections requires the \ccyl part of a cylinder to be even expressed, whereas our old notion of effective connection does not.
Since the \ccyl part is not needed except for explaining the relation of connections and effective connections, I have made it into a separate notion of "contractions".
We might insert some separate references to the literature for contractions and connections.

*** Moved definition of category of strong homotopy equivalences earlier

This is needed I think to make the definition of effective connections nicer (the three natural transformations with greek letters in there made the definition quite ugly to read and were not even used
explicitly afterwards anyway).
Also, I opted for making this an absolute category S_k of strong (left/right) homotopy equivalences, not relative to some functor u : I -> E^to.
First of all, the relativity is not needed here.
But second and more importantly, the relative notion can be recovered quite nicely by just taking the pullback of S_k -> E^to along I -> E^to.
This should improve the section on strong homotopy equivalences (still to be done).