Minor fix.
I forgot to mention that I also swapped the meaning of strong left and right homotopy equivalences.
This makes everything uniform, I hope.
Effective connections require that \lcyl^0 x X is a strong *left* homotopy equivalences.
And f is a strong *left* equivalence if \theta^0 \hatotimes f is a section.