Checked proof of Lemma 7.4.
I do not like the repetitions induced by our notation "f : A -> B exhibits A as a retract of B" (when A and B are larger expressions), but I didn't do anything about it.
I am also not sure whether "exhibit" is the right word here, seeing as this is not a property (the retraction to f is not uniquely determined by f being a section).
Of course, this notation originated with Lemma 3.3.